api/core/plugin/kagents_agent/source_code/requirements.txt
dify_plugin==0.3.3
a2a-sdk==0.2.5
ruamel.yaml==0.18.14